+++ R/AcqOptimizerLbfgsb.R
@@ -141,6 +141,7 @@ AcqOptimizerLbfgsb = R6Class(
       }
 
       y = Inf
+      x = NULL
       n_evals = 0L
       n_restarts = 0L
       while ((n_evals < maxeval || maxeval < 0) && n_restarts <= max_restarts) {
@@ -187,7 +188,8 @@ AcqOptimizerLbfgsb = R6Class(
           res = optimize()
         }
 
-        if (res$objective < y) {
+        # isTRUE guards against a NaN objective (e.g., constant acquisition surface) that would otherwise error here
+        if (isTRUE(res$objective < y)) {
           y = res$objective
           x = res$solution
         }
@@ -197,6 +199,9 @@ AcqOptimizerLbfgsb = R6Class(
         self$state = c(self$state, set_names(list(list(model = res, start = x0)), paste0("iteration_", n_restarts)))
 
         if (restart_strategy == "none") break
+      }
+      if (is.null(x)) {
+        error_acq_optimizer("Acquisition function optimization did not yield a valid solution.")
       }
       as.data.table(as.list(set_names(
         c(x, y * direction),